refactor(gateway): extract chat attachment phase
